Update Lingo re. rename of arches-controlled-lists (#212)
