+++ arches_lingo/hosts.py
@@ -4,5 +4,5 @@ from django_hosts import patterns, host
 host_patterns = patterns(
     "",
     host(re.sub(r"_", r"-", r"arches_lingo"), "arches_lingo.urls", name="arches_lingo"),
-    host(r"arches", "arches_references.urls", name="arches"),
+    host(r"arches", "arches_controlled_lists.urls", name="arches"),
 )
