+++ arches_lingo/settings.py
@@ -238,7 +238,7 @@ INSTALLED_APPS = (
     "django.contrib.staticfiles",
     "django.contrib.gis",
     "django_hosts",
-    "arches_references",
+    "arches_controlled_lists",
     "arches_vue_utils",
     "arches",
     "arches.app.models",
